+++ sales-tax/src/main/java/com/complyt/v1/models/ItemDto.java
@@ -11,17 +11,19 @@ import jakarta.validation.constraints.PositiveOrZero;
 import jakarta.validation.constraints.Size;
 import lombok.With;
 
+import java.math.BigDecimal;
+
 @With
 @Schema(name = "Item")
 public record ItemDto(
-        @PositiveOrZero(message = "Item.unitPrice " + NumericErrorMessages.NOT_NEGATIVE_ERROR) float unitPrice,
-        @PositiveOrZero(message = "Item.quantity " + NumericErrorMessages.NOT_NEGATIVE_ERROR) float quantity,
-        @PositiveOrZero(message = "Item.totalPrice " + NumericErrorMessages.NOT_NEGATIVE_ERROR) float totalPrice,
+        @PositiveOrZero(message = "Item.unitPrice " + NumericErrorMessages.NOT_NEGATIVE_ERROR) BigDecimal unitPrice,
+        @PositiveOrZero(message = "Item.quantity " + NumericErrorMessages.NOT_NEGATIVE_ERROR) BigDecimal quantity,
+        @PositiveOrZero(message = "Item.totalPrice " + NumericErrorMessages.NOT_NEGATIVE_ERROR) BigDecimal totalPrice,
         String description,
         @NotNull(message = "Item.name " + DtoErrorMessages.NOT_NULL_ERROR) @Size(min = 1, max = 256, message = "Item.name " + StringErrorMessages.MINMAX_256_ERROR) String name,
         @NotNull(message = "Item.taxCode " + DtoErrorMessages.NOT_NULL_ERROR) @Size(max = 256, message = "Item.taxCode " + StringErrorMessages.MINMAX_256_ERROR) String taxCode,
         JurisdictionalSalesTaxRulesDto jurisdictionalSalesTaxRules, SalesTaxRatesDto salesTaxRates,
         boolean manualSalesTax,
-        @PositiveOrZero(message = "Item.manualSalesTaxRate " + NumericErrorMessages.NOT_NEGATIVE_ERROR) @DecimalMax(value = "0.2", message = "Item.manualSalesTaxRate" + NumericErrorMessages.DECIMAL_MAX_02_ERROR) float manualSalesTaxRate,
+        @PositiveOrZero(message = "Item.manualSalesTaxRate " + NumericErrorMessages.NOT_NEGATIVE_ERROR) @DecimalMax(value = "0.2", message = "Item.manualSalesTaxRate" + NumericErrorMessages.DECIMAL_MAX_02_ERROR) BigDecimal manualSalesTaxRate,
         TangibleCategoryDto tangibleCategory, TaxableCategoryDto taxableCategory) {
 }
