+++ sales-tax/src/main/java/com/complyt/v1/models/JurisdictionalSalesTaxRulesDto.java
@@ -4,6 +4,7 @@ import com.complyt.domain.sales_tax.product_classification.CalculationType;
 import com.complyt.domain.sales_tax.product_classification.CitySalesTaxRules;
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Schema(name = "JurisdictionalSalesTaxRules")
@@ -11,9 +12,10 @@ public record JurisdictionalSalesTaxRulesDto(
         String name,
         String abbreviation,
         boolean taxable,
-        boolean specialTreatment, CalculationType calculationType,
+        boolean specialTreatment,
+        CalculationType calculationType,
         String description,
-        float calculationValue,
+        BigDecimal calculationValue,
         Map<String, CitySalesTaxRules> cities) {
 
 }
